Make XRPAmount constructor explicit:
Remove the implicit conversion from int64 to XRPAmount. The motivation for this
was noticing that many calls to `to_string` with an integer parameter type were
calling the wrong `to_string` function. Since the calls were not prefixed with
`std::`, and there is no ADL to call `std::to_string`, this was converting the
int to an `XRPAmount` and calling `to_string(XRPAmount)`.

Since `to_string(XRPAmount)` did the same thing as `to_string(int)` this error
went undetected.

+++ src/ripple/protocol/SystemParameters.h
@@ -43,7 +43,7 @@ systemName ()
 static
 constexpr
 XRPAmount
-INITIAL_XRP{ 100'000'000'000 * DROPS_PER_XRP };
+INITIAL_XRP{ static_cast<std::int64_t>(100'000'000'000) * DROPS_PER_XRP };
 
 /** Returns true if the amount does not exceed the initial XRP in existence. */
 inline
